+++ tutor/plugins.py
@@ -1,4 +1,5 @@
 from copy import deepcopy
+import importlib
 import pkg_resources
 
 from . import exceptions
@@ -7,15 +8,14 @@ from . import exceptions
 CONFIG_KEY = "PLUGINS"
 
 
-class Plugins:
+class BasePlugin:
     """
-    Tutor plugins are regular python packages that have a 'tutor.plugin.v0' entrypoint.
-
-    The API for Tutor plugins is currently in development. The entrypoint will switch to
-    'tutor.plugin.v1' once it is stabilised.
-
-    This entrypoint must point to a module or a class that implements one or more of the
+    Tutor plugins are defined by a name and an object that implements one or more of the
     following properties:
+
+    `config` (dict str->dict(str->str)): contains "add", "set", "default" keys. Entries
+    in these dicts will be added or override the global configuration. Keys in "add" and
+    "set" will be prefixed by the plugin name in uppercase.
 
     `patches` (dict str->str): entries in this dict will be used to patch the rendered
     Tutor templates. For instance, to add "somecontent" to a template that includes '{{
@@ -36,11 +36,94 @@ class Plugins:
 
     It is then assumed that there are `myplugin/hooks/service1/init` and
     `myplugin/hooks/service2/init` templates in the plugin `templates` directory.
+
+    `command` (click.Command): if a plugin exposes a `command` attribute, users will be able to run it from the command line as `tutor pluginname`.
+    """
+
+    def __init__(self, name, obj):
+        self.name = name
+        self.config = get_callable_attr(obj, "config", {})
+        self.patches = get_callable_attr(obj, "patches", default={})
+        self.hooks = get_callable_attr(obj, "hooks", default={})
+        self.templates_root = get_callable_attr(obj, "templates", default=None)
+        self.command = getattr(obj, "command", None)
+
+    def config_key(self, key):
+        """
+        Config keys in the "add" and "defaults" dicts should be prefixed by the plugin name, in uppercase.
+        """
+        return self.name.upper() + "_" + key
+
+    @property
+    def config_add(self):
+        return self.config.get("add", {})
+
+    @property
+    def config_set(self):
+        return self.config.get("set", {})
+
+    @property
+    def config_defaults(self):
+        return self.config.get("defaults", {})
+
+    @property
+    def version(self):
+        raise NotImplementedError
+
+    @classmethod
+    def iter_installed(cls):
+        raise NotImplementedError
+
+
+class EntrypointPlugin(BasePlugin):
+    """
+    Entrypoint plugins are regular python packages that have a 'tutor.plugin.v0' entrypoint.
+
+    The API for Tutor plugins is currently in development. The entrypoint will switch to
+    'tutor.plugin.v1' once it is stabilised.
     """
 
     ENTRYPOINT = "tutor.plugin.v0"
+
+    def __init__(self, entrypoint):
+        super().__init__(entrypoint.name, entrypoint.load())
+        self.entrypoint = entrypoint
+
+    @property
+    def version(self):
+        return self.entrypoint.dist.version
+
+    @classmethod
+    def iter_installed(cls):
+        for entrypoint in pkg_resources.iter_entry_points(cls.ENTRYPOINT):
+            yield cls(entrypoint)
+
+
+class OfficialPlugin(BasePlugin):
+    """
+    Official plugins have a "plugin" module which exposes a __version__
+    attribute.
+    Official plugins should be manually added to INSTALLED.
+    """
+
+    INSTALLED = []
+
+    def __init__(self, name):
+        self.module = importlib.import_module("tutor{}.plugin".format(name))
+        super().__init__(name, self.module)
+
+    @property
+    def version(self):
+        return self.module.__version__
+
+    @classmethod
+    def iter_installed(cls):
+        yield from cls.INSTALLED
+
+
+class Plugins:
+
     INSTANCE = None
-    EXTRA_INSTALLED = {}
 
     def __init__(self, config):
         self.config = deepcopy(config)
@@ -48,27 +131,24 @@ class Plugins:
         self.hooks = {}
         self.template_roots = {}
 
-        for plugin_name, plugin in self.iter_enabled():
-            patches = get_callable_attr(plugin, "patches", {})
-            for patch_name, content in patches.items():
+        for plugin in self.iter_enabled():
+            for patch_name, content in plugin.patches.items():
                 if patch_name not in self.patches:
                     self.patches[patch_name] = {}
-                self.patches[patch_name][plugin_name] = content
+                self.patches[patch_name][plugin.name] = content
 
-            hooks = get_callable_attr(plugin, "hooks", {})
-            for hook_name, services in hooks.items():
+            for hook_name, services in plugin.hooks.items():
                 if hook_name not in self.hooks:
                     self.hooks[hook_name] = {}
-                self.hooks[hook_name][plugin_name] = services
+                self.hooks[hook_name][plugin.name] = services
 
-            templates_root = get_callable_attr(plugin, "templates")
-            if templates_root:
-                self.template_roots[plugin_name] = templates_root
+            if plugin.templates_root:
+                self.template_roots[plugin.name] = plugin.templates_root
 
     @classmethod
     def clear(cls):
         cls.INSTANCE = None
-        cls.EXTRA_INSTALLED.clear()
+        OfficialPlugin.INSTALLED.clear()
 
     @classmethod
     def instance(cls, config):
@@ -78,20 +158,21 @@ class Plugins:
 
     @classmethod
     def iter_installed(cls):
-        yield from cls.EXTRA_INSTALLED.items()
-        for name, module in cls.iter_installed_entrypoints():
-            if name not in cls.EXTRA_INSTALLED:
-                yield name, module
-
-    @classmethod
-    def iter_installed_entrypoints(cls):
-        for entrypoint in pkg_resources.iter_entry_points(cls.ENTRYPOINT):
-            yield (entrypoint.name, entrypoint.load())
+        """
+        Iterate on all installed plugins. Plugins are deduplicated by name.
+        """
+        classes = [OfficialPlugin, EntrypointPlugin]
+        installed_plugin_names = set()
+        for PluginClass in classes:
+            for plugin in PluginClass.iter_installed():
+                if plugin.name not in installed_plugin_names:
+                    installed_plugin_names.add(plugin.name)
+                    yield plugin
 
     def iter_enabled(self):
-        for name, plugin in self.iter_installed():
-            if is_enabled(self.config, name):
-                yield name, plugin
+        for plugin in self.iter_installed():
+            if is_enabled(self.config, plugin.name):
+                yield plugin
 
     def iter_patches(self, name):
         plugin_patches = self.patches.get(name, {})
@@ -114,7 +195,7 @@ def get_callable_attr(plugin, attr_name, default=None):
 
 
 def is_installed(name):
-    plugin_names = [name for name, _ in iter_installed()]
+    plugin_names = [plugin.name for plugin in iter_installed()]
     return name in plugin_names
 
 
@@ -156,13 +237,3 @@ def iter_hooks(config, hook_name):
 
 def iter_template_roots(config):
     yield from Plugins.instance(config).iter_template_roots()
-
-
-def iter_commands():
-    """
-    Iterate over all plugins that provide a `command` attribute.
-    """
-    for plugin_name, plugin in iter_installed():
-        command = getattr(plugin, "command", None)
-        if command:
-            yield plugin_name, command