+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 Note: Breaking changes between versions are indicated by "💥".
 
+## Unreleased
+
+- 💥[Improvement] The output of `plugins list` now includes plugin version numbers.
+
 ## 3.10.1 (2020-01-13)
 
 - [Improvement] Explicitely point to docker.io images, when necessary, for [podman](https://podman.io/) compatibility.